backend/requirements.txt
Flask==2.3.3
flask-pymongo==2.3.0
flask-bcrypt==1.0.0
python-dotenv>=0.19.2
requests==2.32.3   
flask-cors==5.0.0